move genomic_interval_index into fusion
